+++ compiler-core/src/analyse.rs
@@ -229,6 +229,7 @@ pub fn infer_module<A>(
         module_types: types,
         module_types_constructors: types_constructors,
         module_values: values,
+        todo_encountered: contains_todo,
         accessors,
         ..
     } = env;
@@ -244,8 +245,9 @@ pub fn infer_module<A>(
             values,
             accessors,
             origin,
-            unused_imports,
             package: package.clone(),
+            unused_imports,
+            contains_todo,
         },
     })
 }
@@ -950,6 +952,7 @@ fn infer_module_constant(
         annotation,
         public,
         value,
+        deprecation,
         ..
     } = c;
 
@@ -960,7 +963,7 @@ fn infer_module_constant(
 
     let variant = ValueConstructor {
         public,
-        deprecation: Deprecation::NotDeprecated,
+        deprecation: deprecation.clone(),
         variant: ValueConstructorVariant::ModuleConstant {
             documentation: doc.clone(),
             location,
@@ -993,6 +996,7 @@ fn infer_module_constant(
         value: Box::new(typed_expr),
         type_,
         supported_targets,
+        deprecation,
     }))
 }
 
@@ -1074,6 +1078,7 @@ fn generalise_module_constant(
         value,
         type_,
         supported_targets,
+        deprecation,
     } = constant;
     let typ = type_.clone();
     let type_ = type_::generalise(typ);
@@ -1089,7 +1094,7 @@ fn generalise_module_constant(
         variant.clone(),
         type_.clone(),
         public,
-        Deprecation::NotDeprecated,
+        deprecation.clone(),
     );
 
     environment.insert_module_value(
@@ -1097,7 +1102,7 @@ fn generalise_module_constant(
         ValueConstructor {
             public,
             variant,
-            deprecation: Deprecation::NotDeprecated,
+            deprecation: deprecation.clone(),
             type_: type_.clone(),
         },
     );
@@ -1111,6 +1116,7 @@ fn generalise_module_constant(
         value,
         type_,
         supported_targets,
+        deprecation,
     })
 }