+++ compiler-core/src/call_graph/into_dependency_order_tests.rs
@@ -53,6 +53,7 @@ fn parse_and_order(
                 value: Box::from(const_value),
                 supported_targets: SupportedTargets::all(),
                 type_: (),
+                deprecation: Deprecation::NotDeprecated,
             }
         })
         .collect_vec();