+++ compiler-core/src/metadata/tests.rs
@@ -28,6 +28,7 @@ fn roundtrip(input: &ModuleInterface) -> ModuleInterface {
 
 fn constant_module(constant: TypedConstant) -> ModuleInterface {
     ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a".into(),
@@ -72,6 +73,7 @@ fn bit_array_segment_option_module(option: TypedConstantBitArraySegmentOption) -
 #[test]
 fn empty_module() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "one/two".into(),
@@ -87,6 +89,7 @@ fn empty_module() {
 #[test]
 fn module_with_private_type() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a/b".into(),
@@ -113,6 +116,7 @@ fn module_with_private_type() {
 #[test]
 fn module_with_unused_import() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a".into(),
@@ -131,6 +135,7 @@ fn module_with_unused_import() {
 #[test]
 fn module_with_app_type() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a/b".into(),
@@ -157,6 +162,7 @@ fn module_with_app_type() {
 #[test]
 fn module_with_fn_type() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a/b".into(),
@@ -183,6 +189,7 @@ fn module_with_fn_type() {
 #[test]
 fn module_with_tuple_type() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a/b".into(),
@@ -215,6 +222,7 @@ fn module_with_generic_type() {
 
     fn make(t1: Arc<Type>, t2: Arc<Type>) -> ModuleInterface {
         ModuleInterface {
+            contains_todo: false,
             package: "some_package".into(),
             origin: Origin::Src,
             name: "a/b".into(),
@@ -247,6 +255,7 @@ fn module_with_type_links() {
 
     fn make(type_: Arc<Type>) -> ModuleInterface {
         ModuleInterface {
+            contains_todo: false,
             package: "some_package".into(),
             origin: Origin::Src,
             name: "a".into(),
@@ -275,6 +284,7 @@ fn module_with_type_links() {
 #[test]
 fn module_type_to_constructors_mapping() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a".into(),
@@ -298,6 +308,7 @@ fn module_type_to_constructors_mapping() {
 #[test]
 fn module_fn_value() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a".into(),
@@ -333,6 +344,7 @@ fn module_fn_value() {
 #[test]
 fn deprecated_module_fn_value() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a".into(),
@@ -370,6 +382,7 @@ fn deprecated_module_fn_value() {
 #[test]
 fn private_module_fn_value() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a".into(),
@@ -407,6 +420,7 @@ fn private_module_fn_value() {
 #[test]
 fn module_fn_value_regression() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a/b/c".into(),
@@ -443,6 +457,7 @@ fn module_fn_value_regression() {
 #[test]
 fn module_fn_value_with_field_map() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a".into(),
@@ -481,6 +496,7 @@ fn record_value() {
     let mut random = rand::thread_rng();
 
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a".into(),
@@ -520,6 +536,7 @@ fn record_value_with_field_map() {
     let mut random = rand::thread_rng();
 
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a".into(),
@@ -560,6 +577,7 @@ fn record_value_with_field_map() {
 #[test]
 fn accessors() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a".into(),
@@ -764,6 +782,7 @@ fn constant_var() {
     };
 
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a".into(),
@@ -968,6 +987,7 @@ fn constant_bit_array_native() {
 #[test]
 fn deprecated_type() {
     let module = ModuleInterface {
+        contains_todo: false,
         package: "some_package".into(),
         origin: Origin::Src,
         name: "a/b".into(),
@@ -985,6 +1005,23 @@ fn deprecated_type() {
             },
         )]
         .into(),
+        types_value_constructors: HashMap::new(),
+        values: HashMap::new(),
+        unused_imports: Vec::new(),
+        accessors: HashMap::new(),
+    };
+    assert_eq!(roundtrip(&module), module);
+}
+
+#[test]
+fn contains_todo() {
+    let module = ModuleInterface {
+        contains_todo: true,
+        //             ^^^^ It does, it does!
+        package: "some_package".into(),
+        origin: Origin::Src,
+        name: "a/b".into(),
+        types: [].into(),
         types_value_constructors: HashMap::new(),
         values: HashMap::new(),
         unused_imports: Vec::new(),