+++ CHANGELOG.md
@@ -2,11 +2,29 @@
 
 ## Unreleased
 
-### Build tool
+### Build tool changes
 
 - The search bar in generated docs now has a darker background color.
 
-## v0.34.0-rc3 - 2023-01-12
+- If a package contains a `todo` expression then the build tool will now refuse
+  to publish it to Hex.
+
+
+## v0.34.1 - 2023-01-17
+
+### Build tool changes
+
+- Support has been added for using SourceHut as a repository.
+
+### Bug fixes
+
+- Fixed a bug where long function headers with external implementations could
+  format incorrectly.
+- The `@deprecated` attribute can now be used to annotate module constants.
+  This will cause a warning to be emitted when the constant is used.
+
+
+## v0.34.0 - 2023-01-16
 
 ### Language changes
 
@@ -15,6 +33,11 @@
   well as external implementations for both targets. This is because having a
   default Gleam implementation means the code is future-proof and continues to
   be cross platform even if a new target is added.
+
+### Bug fixes
+
+- Fixed a bug where function heads would go over the line limit in the
+  formatter.
 
 
 ## v0.34.0-rc2 - 2023-01-11
@@ -28,7 +51,7 @@
   target.
 
 
-## v0.34.0-rc1 - 2024-02-07
+## v0.34.0-rc1 - 2024-01-07
 
 ### Language changes